refactor: Remove compliance prediction files and move others
