+++ .pre-commit-config.yaml
@@ -23,7 +23,7 @@ repos:
     stages: [commit]
   - id: lint-check
     name: Linting (ruff)
-    entry: ruff check
+    entry: ruff check --fix
     language: system
     types: [python]
     stages: [commit]
